moved all /pkg kube stuff to /pkg/platform/kube
