+++ cmd/controller/app/controller.go
@@ -20,10 +20,10 @@ import (
 	"fmt"
 	"strconv"
 
-	"github.com/nuclio/nuclio/pkg/controller"
+	"github.com/nuclio/nuclio/pkg/platform/kube/controller"
 	"github.com/nuclio/nuclio/pkg/errors"
-	"github.com/nuclio/nuclio/pkg/functioncr"
-	"github.com/nuclio/nuclio/pkg/functiondep"
+	"github.com/nuclio/nuclio/pkg/platform/kube/functioncr"
+	"github.com/nuclio/nuclio/pkg/platform/kube/functiondep"
 	"github.com/nuclio/nuclio/pkg/zap"
 
 	"github.com/nuclio/nuclio-sdk"
