+++ cmd/controller/app/types.go
@@ -17,7 +17,7 @@ limitations under the License.
 package app
 
 import (
-	"github.com/nuclio/nuclio/pkg/functioncr"
+	"github.com/nuclio/nuclio/pkg/platform/kube/functioncr"
 
 	v1beta1 "k8s.io/api/apps/v1beta1"
 	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
