+++ pkg/platform/kube/consumer.go
@@ -18,8 +18,8 @@ package kube
 
 import (
 	"github.com/nuclio/nuclio/pkg/errors"
-	"github.com/nuclio/nuclio/pkg/functioncr"
-	"github.com/nuclio/nuclio/pkg/functiondep"
+	"github.com/nuclio/nuclio/pkg/platform/kube/functioncr"
+	"github.com/nuclio/nuclio/pkg/platform/kube/functiondep"
 
 	"github.com/nuclio/nuclio-sdk"
 	"k8s.io/client-go/kubernetes"
