+++ pkg/platform/kube/functioncr/client.go
