+++ pkg/platform/kube/functioncr/condition.go
