+++ pkg/platform/kube/functioncr/function_test.go
