+++ pkg/platform/kube/functioncr/watcher.go
