+++ pkg/platform/kube/functiondep/client.go
@@ -23,7 +23,7 @@ import (
 	"time"
 
 	"github.com/nuclio/nuclio/pkg/errors"
-	"github.com/nuclio/nuclio/pkg/functioncr"
+	"github.com/nuclio/nuclio/pkg/platform/kube/functioncr"
 
 	"github.com/nuclio/nuclio-sdk"
 	v1beta1 "k8s.io/api/apps/v1beta1"
