+++ pkg/platform/kube/functiondep/client_test.go
@@ -19,7 +19,7 @@ package functiondep
 import (
 	"testing"
 
-	"github.com/nuclio/nuclio/pkg/functioncr"
+	"github.com/nuclio/nuclio/pkg/platform/kube/functioncr"
 	"github.com/nuclio/nuclio/pkg/zap"
 
 	"github.com/nuclio/nuclio-sdk"
